Move validator to core
It is useful for both the api and the adapters.
